Move Devo/HA to talents folder
